app_server/requirements.txt
alabaster==0.7.8
alembic==0.8.2
amqp==1.4.9
aniso8601==1.1.0
anyjson==0.3.3
apiclient==1.0.2
astroid==1.4.5
Babel==2.3.4
bcrypt==2.0.0
beat==0.1.6
beautifulsoup4==4.4.1
billiard==3.3.0.23
blinker==1.4
boto==2.38.0
celery==3.1.23
cffi==1.4.2
colorama==0.3.7
connexion==1.0.67
coverage==4.0.3
decorator==4.0.2
discovery==0.0.4
docutils==0.12
eventlet==0.19.0
extras==1.0.0
fixtures==3.0.0
flasgger==0.5.12
Flask==0.10.1
Flask-Assets==0.11
Flask-Bcrypt==0.6.2
Flask-Cache==0.13.1
Flask-Celery==2.4.3
Flask-Cors==2.1.2
flask-csrf==0.9.2
Flask-HTTPAuth==2.6.0
flask-log==0.1.0
Flask-Logging==0.1.3
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.6.0
Flask-Principal==0.4.0
Flask-Redis==0.1.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-S3==0.1.7
Flask-Script==0.6.3
Flask-Security==1.7.4
Flask-Sessions==0.1.5
Flask-SocketIO==2.2
Flask-SQLAlchemy==1.0
flask-swagger==0.2.12
Flask-Uploads==0.1.3
Flask-WTF==0.12
funcsigs==0.4
gcm==0.3
gevent==1.1.1
google-api-python-client==1.5.1
greenlet==0.4.10
gunicorn==19.4.5
httpauth==0.2
httplib2==0.9.2
imagesize==0.7.1
iso8601==0.1.11
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
kombu==3.0.35
lazy-object-proxy==1.2.2
linecache2==1.0.0
locustio==0.7.3
Mako==1.0.3
MarkupSafe==0.23
migrate==0.3.7
mistune==0.7.2
mock==1.3.0
modulegraph==0.10.4
msgpack-python==0.4.7
Naked==0.1.31
nose==1.3.7
oauth2client==2.2.0
passlib==1.6.5
pathlib==1.0.1
pbr==1.8.1
Pillow==3.3.0
pluggy==0.3.1
psycopg2==2.6.1
py==1.4.31
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
Pygments==2.1.3
PyJWT==1.4.0
pylint==1.5.5
pyparsing==2.0.1
python-bcrypt==0.3.1
python-dateutil==2.4.2
python-editor==0.5
python-engineio==0.8.6
python-json-logger==0.1.4
python-logstash==0.4.5
python-mimeparse==1.5.2
python-openid==2.2.5
python-socketio==1.1
pytz==2016.3
PyYAML==3.11
redis==2.10.5
requests==2.5.0
rsa==3.4.2
shellescape==3.4.1
simplejson==3.8.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.2
sphinxcontrib-httpdomain==1.5.0
SQLAlchemy==1.0.12
strict-rfc3339==0.7
testing.common.database==1.1.0
testing.redis==1.1.0
testtools==2.2.0
tox==2.3.1
traceback2==1.4.0
Twisted==16.3.0
unittest2==1.1.0
uritemplate==0.6
urllib3==1.15.1
virtualenv==15.0.2
webassets==0.11.1
Werkzeug==0.11.2
wrapt==1.10.8
WTForms==1.0.5
zope.interface==4.2.0
